+++ src/section/templates.tsx
@@ -1,7 +1,8 @@
 export default function TemplatesSection(){
      return (
-          <div className="w-full h-[50dvh] py-4 px-8">
-               <div className="w-full max-w-360">
+          <div className="w-full h-[75dvh] py-4 px-8 flex justify-center items-center">
+               <div className="w-full max-w-360 flex justify-center items-center flex-col gap-4">
+                    <h2 className="text-xl md:text-2xl lg:text-3xl font-semibold border-b border-primary pb-1 mb-4 w-fit">Շաբլոններ</h2>
                     TODO: Make an Templates section
                </div>
           </div>