+++ changelog.md
@@ -10,3 +10,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Header
 - Footer
 - Hero Section
+- About Section
+- Search Popup UI